+++ convex/_generated/api.d.ts
@@ -10,8 +10,10 @@
 
 import type * as appointments from "../appointments.js";
 import type * as auth from "../auth.js";
+import type * as crons from "../crons.js";
 import type * as email from "../email.js";
 import type * as emailOutbox from "../emailOutbox.js";
+import type * as emailOutboxActions from "../emailOutboxActions.js";
 import type * as employees from "../employees.js";
 import type * as http from "../http.js";
 import type * as lib_auth from "../lib/auth.js";
@@ -31,8 +33,10 @@ import type {
 declare const fullApi: ApiFromModules<{
   appointments: typeof appointments;
   auth: typeof auth;
+  crons: typeof crons;
   email: typeof email;
   emailOutbox: typeof emailOutbox;
+  emailOutboxActions: typeof emailOutboxActions;
   employees: typeof employees;
   http: typeof http;
   "lib/auth": typeof lib_auth;
